+++ abeja/training/__init__.py
@@ -1,6 +1,7 @@
 # flake8: noqa
 from .client import Client
 from .common import SizedIterable
-from .job_definition import JobDefinitions, JobDefinition, JobDefinitionVersions, JobDefinitionVersion, Job
+from .job_definition import JobDefinitions, JobDefinition, Job
+from .job_definition_version import JobDefinitionVersion, JobDefinitionVersions
 from .api.client import APIClient
 from .statistics import Statistics
